requirements.txt
# Core FastAPI and Web server
fastapi[all]

# Database
SQLAlchemy
psycopg2-binary
alembic

# Authentication
passlib==1.7.4          # ← バージョン固定（bcrypt extrasは不要）
bcrypt==4.0.1
python-jose[cryptography]
pyasn1>=0.4.1,<0.5.0

# Other utilities
python-multipart
Jinja2
python-dotenv
Pillow

# AI SDKs
openai
anthropic
google-generativeai
cohere
perplexipy
deepl
requests
httpx
replicate
stability-sdk
# assemblyai  # disabled: talk mode removed

# Dev / Test
pytest==8.2.2        # 例：インストールされたバージョンを固定
playwright